Rudimentary error logging over bluetooth
The logging library can be used to send messages over bluetooth; but
long messages tend to cause bugs / lost data from other things being
transmitted.
As a workaround, I have implemented some error codes that are sent the
same way as component values. This is an early work in progress, so the
format and/or error codes may change