Initialize Controller in setup(); add 8 valves
The initialization thing is to correct a really weird bug which appeared after updating the ESP32 i2c libraries. Wire.begin() was called in Controller's constructor, which was called before setup(). That didn't throw any errors, but resulted in instructions to the PCA9896 having no effect.

Moving the initialization to be called in setup() solves this.