Transmit pressure regulator setpoints
This implementation is a duct-tape solution at best. Ideally, a big
rewrite of the communication protocol should be done, where we don't
rely on hardcoded constants for each possible component. But that's
not possible immediately, so this will work in the meantime.

As with any update to constants.h, this requires a corresponding
update of the host software.