Fix bug where multiple commands in a row were ignored
When sending a multiplexer command, nothing happened. The line added is
on the PC side of the code, but I had forgotten to add it here. Without
the break statement added, when a buffer contains multiple commands, the
whole thing ends up being parsed with all commands concatenated
together. Resulting in none of them being correctly interpreted.